requirements.txt
websocket-client==1.6.1
keyboard==0.13.5
evdev==1.6.1
